+++ src/presentation/controllers/waypointController.js
@@ -4,7 +4,7 @@ const { sendSuccess } = require('../../utils/httpResponse');
 const createWaypoint = async (req, res, next) => {
   try {
     const waypointData = req.body;
-    const newWaypoint = await waypointService.createWaypointService(waypointData);
+    const newWaypoint = await waypointService.createWaypointService(waypointData, req.user.userId);
     return sendSuccess(res, 201, newWaypoint);
   } catch (error) {
     next(error);
@@ -14,7 +14,7 @@ const createWaypoint = async (req, res, next) => {
 const getWaypoint = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const waypoint = await waypointService.getWaypointService(id);
+    const waypoint = await waypointService.getWaypointService(id, req.user.userId);
     return sendSuccess(res, 200, waypoint);
   } catch (error) {
     next(error);
@@ -24,7 +24,7 @@ const getWaypoint = async (req, res, next) => {
 const getAllWaypoints = async (req, res, next) => {
   try {
     const { mapId } = req.query;
-    const waypoints = await waypointService.getAllWaypointsService(mapId);
+    const waypoints = await waypointService.getAllWaypointsService(mapId, req.user.userId);
     return sendSuccess(res, 200, waypoints);
   } catch (error) {
     next(error);
@@ -35,7 +35,7 @@ const updateWaypoint = async (req, res, next) => {
   try {
     const { id } = req.params;
     const updateData = req.body;
-    const updatedWaypoint = await waypointService.updateWaypointService(id, updateData);
+    const updatedWaypoint = await waypointService.updateWaypointService(id, updateData, req.user.userId);
     return sendSuccess(res, 200, updatedWaypoint);
   } catch (error) {
     next(error);
@@ -45,7 +45,7 @@ const updateWaypoint = async (req, res, next) => {
 const deleteWaypoint = async (req, res, next) => {
   try {
     const { id } = req.params;
-    await waypointService.deleteWaypointService(id);
+    await waypointService.deleteWaypointService(id, req.user.userId);
     return res.status(204).send();
   } catch (error) {
     next(error);